Move Quick-Pick Treatments logic out of HomeStateModel
Splits the suggestion scoring/fetching helpers and the load/store/enact
methods into HomeStateModel+QuickPickTreatments.swift, following the
existing HomeStateModel+CGM.swift extension pattern, to keep
HomeStateModel from growing further.
